+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added compatibility with `nikic/php-parser` `5.x`
 - Removed compatibility with `nikic/php-parser` `4.x`
 
+💥 **BC breaks**
+
+- deprecated constant `VarExporter::INLINE_NUMERIC_SCALAR_ARRAY` has been removed, please use `INLINE_SCALAR_LIST` instead
+
 ## [0.4.0](https://github.com/brick/varexporter/releases/tag/0.4.0) - 2023-09-01
 
 Minimum PHP version is now `7.4`. No breaking changes.